# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: string;
  username: string;
  name: string;
  timezone: string;
  createdAt: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 3
     * @maxLength 40
     */
  username?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  timezone?: string;
}

export interface EventType {
  id: number;
  userId: string;
  title: string;
  slug: string;
  /** @nullable */
  description: string | null;
  durationMinutes: number;
  color: string;
  isActive: boolean;
  /** If false, the event type is secret — bookable via direct link but not listed on the public profile page */
  isPublic: boolean;
  createdAt: string;
}

export interface EventTypeInput {
  /** @minLength 1 */
  title: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minimum 5
     * @maximum 480
     */
  durationMinutes: number;
  color?: string;
  isActive?: boolean;
  isPublic?: boolean;
}

export interface EventTypeUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minimum 5
     * @maximum 480
     */
  durationMinutes?: number;
  color?: string;
  isActive?: boolean;
  isPublic?: boolean;
}

export interface AvailabilityRule {
  id: number;
  /**
     * 0 = Sunday .. 6 = Saturday
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  /** 24h HH:mm in the host's timezone */
  startTime: string;
  /** 24h HH:mm in the host's timezone */
  endTime: string;
}

export interface AvailabilityRuleInput {
  /**
     * @minimum 0
     * @maximum 6
     */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
}

export interface AvailabilityRulesInput {
  rules: AvailabilityRuleInput[];
}

export type BookingStatusFilter = typeof BookingStatusFilter[keyof typeof BookingStatusFilter];


export const BookingStatusFilter = {
  upcoming: 'upcoming',
  past: 'past',
  cancelled: 'cancelled',
  all: 'all',
} as const;

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  eventTypeId: number;
  eventTypeTitle: string;
  hostUserId: string;
  inviteeName: string;
  inviteeEmail: string;
  inviteeTimezone?: string;
  /** @nullable */
  notes?: string | null;
  startTime: string;
  endTime: string;
  status: BookingStatus;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  inviteeName: string;
  inviteeEmail: string;
  /** @nullable */
  notes?: string | null;
  /** UTC ISO instant for the slot start */
  startTime: string;
  /** IANA timezone the invitee is booking from */
  timezone: string;
}

export interface DashboardSummary {
  totalEventTypes: number;
  upcomingBookings: number;
  bookingsToday: number;
  bookingsThisWeek: number;
}

export interface PublicEventType {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  description: string | null;
  durationMinutes: number;
  color: string;
  hostName: string;
  hostUsername: string;
  hostTimezone: string;
}

export interface PublicProfile {
  username: string;
  name: string;
  timezone: string;
  eventTypes: PublicEventType[];
}

export interface SlotDay {
  /** YYYY-MM-DD in the requested timezone */
  date: string;
  slots: string[];
}

export type ListMyBookingsParams = {
status?: BookingStatusFilter;
};

export type GetPublicEventTypeSlotsParams = {
username: string;
slug: string;
/**
 * Inclusive start date (YYYY-MM-DD)
 */
startDate: string;
/**
 * Inclusive end date (YYYY-MM-DD)
 */
endDate: string;
/**
 * IANA timezone the caller wants slots expressed in
 */
timezone: string;
};