# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  eventTypeId: number;
  eventTypeTitle: string;
  hostUserId: string;
  inviteeName: string;
  inviteeEmail: string;
  inviteeTimezone?: string;
  /** @nullable */
  notes?: string | null;
  startTime: Date;
  endTime: Date;
  status: BookingStatus;
  createdAt: Date;
}