# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 1 */
  inviteeName: string;
  inviteeEmail: string;
  /** @nullable */
  notes?: string | null;
  /** UTC ISO instant for the slot start */
  startTime: Date;
  /** IANA timezone the invitee is booking from */
  timezone: string;
}