# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  username: string;
  name: string;
  timezone: string;
  createdAt: Date;
}